Add support groups to products
The idea is to automatically add user who buy the product to a dedicated
group. This is not coded in the present commit yet. Only a server action
is defined here to easily accomplish this.

Also note that there is no mecanism to automatically remove users
automatically from the above-mentionned groups. This has to be done
manually if required (in case of support group change, sale cancellation
or the like).