Turn product rental_price into a ratio of the lease amount to the product price
This fixes an issue regarding automatic variant price changes in the
e-commerce product page.

This ratio is used to compute the rental price dynamically according to
variant extra prices when users choose them.

The DOM elements that display the extra price of the variants are simply
hidden.

This whole things is supposed to be temporary as it is a hackish way to
handle the difference between the rental price and the product price
(the initial deposit). The variant extra prices need to be coherent with
the product price, not the rental price, which is counter-intuitive.