Fix active_id name error when browsing directly to a sale opportunity
See #891.

We need to trick Odoo by returning an action json description with the
id (and xml_id) of the server action that computes the view to be used
for present lead according to the custom_view setting of the team, so
that the direct link calls our code. As our code needs the active_id to
be set to the team id, we add it in the URL parameters for the direct
browsing use case.