Add payment.acquirer methods to ease slimpay mandates dump and restore
The former uses Slimpay's API to get a description of all signed
mandates and serialize these into a json file. The file will typically
be backed-up and made available to a test environment restore script.

The restore script can use the second method to restore the signed
mandates into a Slimpay sandbox. Only the unknown mandates will be
recreated in the sandbox.