Fix wrong product code for international returns despite a correct test
The code determining the product code was based on the sender country
not being 'FR', but it was False in the test; fixing this broke the
tests however because they call a low-level function (`shipping_data`)
that was not in charge of swapping the sender and the recipient in case
of a return: the higher level method Parcel.colissimo_label was. As a
consequence, the tested country was always the sender's. This is now
fixed and the `shipping_data` function is now suitable for testing in
itself.