Factor out to present module the code from the commown that adds risk analysis related fields to crm_lead and crm_team
CRM team attribute used_for_risk_analysis controls the display of those
fields on a CRM lead. A related field on the lead is used to relay the
team's used_for_risk_analysis field and used to control the visibility
of the lead's dedicated attributes.

This replaces the old mecanism from the commown module which used to
work with different views, activated from the click on the team on the
dashboard. This mecanism was indeed problematic because different user
paths led to different views of the same lead.