[tests] Make lead naming test relative to coupon more realistic
Use the sale.order `action_confirm` API instead of changing the order state directly.
This triggers a bug relative to this API override in the website_sale_coupon module where
the actions are called before the coupons are validated, leading to a wrong sale order name
(when coupons are implied).