[IMP] Introduce proper API to choose if a payment issue should be handled automatically
Remove the previous convention on Slimpay payment reference starting
with 'TR' (abondonned in the payment_slimpay module) to find the invoice
related to a Slimpay payment issue and handle it automatically. The
invoice is now always linked to the corresponding Odoo issue as soon as
an Odoo transaction with an `acquirer_reference` attribute equal to
Slimpay's payment reference exists and references an invoice (via the
`reference` attribute).
Also the decision to handle the issue automatically is delegated to a
specific method, `slimpay_payment_issue_process_automatically` which
returns a boolean and can easily be overriden to customize the "handle
them all automatically" behaviourof current module.