Execute grouped payment retrials in individual jobs to avoid global rollbacks
Before that, an exception for a payment retrial in a serie led to the
payment retrials in success before the crash to be reexecuted again and
again.

Tests are not perfectly accurate in that matter as we use savepoints to
simulate separate jobs although they use different transactions in real
life. As a consequence, all DB changes in a job with an exception are
rollbacked, though IRL they are not and the exception is visible in the
DB.