Contract template views: tree for lines is no more editable, form view added
Discount formula edition requires a large widget which a tree view cannot offer.
As a consequence, tree view is no more editable and a form view is added for the
contract lines.

Use AceEditor as a widget for discount formula. Improvement to be done: use a
yaml mode for Ace.