Add a lead picking creation wizard
This will fix a bug that occured sometimes when the pack operation location
was wrong. This can happen in v10 if the picking's location is a view which
has several children location: the chosen pack op location is sometimes a
location which does not hold the user-chosen lot.

The implementation removes quite some code on contract and related test.