[commown_contract_variable_discount] Add a coupon-driven discount condition
The discount condition is True when the contract is related to a sale
order associated to a coupon from the campaign specified in the
discount.

Note that this mecanism has only been implemented at contract template
level: it is currently not possible to create such a discount specific
to a contract.