Fix immediately rejected transactions' acquirer reference
This kind of immediate reject only happens on slimpay testing environments.
The only change here is that the slimpay reference is now also stored when the payment is rejected.
This is important for automatic slimpay issue handling tests: without this, unpaid invoices cannot be retrieved as the transaction reference is the link between slimpay and odoo.