Employees previously needed to be in stock manager group to use res.partner set_customer_location method
This method is called very often (for instance when linking a
project.issue with a contract), so it must be callable by anyone: the
creation of a stock location for a customer is a normal operation at
Commown.

We thus use sudo() to make sure the stock location creation does not
prevent employees to do their normal job.