Do not optout at a contract end if the discount was not actually valid
This fixes a bug where a customer has 2 contracts, one with a valid campaign,
the other without. Ending that last contract before the fix used to optout
the customer from the valid campaign of the other contract.

Also move optin to the discount line `compute` method so that it is not
called when checking coupon validity, for instance at the end of a contract.