Do not optout at a contract end if the discount was not actually valid
This fixes a bug where a customer has 2 contracts, one with a valid campaign,
the other without. Ending that last contract before the fix used to optout
the customer from the valid campaign of the other contract.

Also move optin to the discount line `compute` method so that it is not
called when checking coupon validity, for instance at the end of a contract.

Tests have changed for several reasons:
- the optin http API is less called, so request mocks checks changed
- the l10n dependency was added to have a stable chart of accounting,
  so the prices change as a consequence of the new taxes
- a requests_mock change (?) implied a change to the way a HTTP return
  code of 422 is mocked (see double optin test)