Fix payment_slimpay_issue tests in the CI
The reason for the crash was that in the CI, the commown module is
installed, overriding the project.task's
slimpay_payment_issue_process_automatically method. We want to test
current module's implementation and do not want to rewrite the tests
with contract-generated invoices (required by the commown module's
implementation), as it is slow and cumbersome.

The chosen solution was so to monkey patch the commown's implementation
when the commown module is installed (i.e. importable) and in the
present tests only.

Another error had the same source: commown module being installed in
the CI context. In this case, a SMS was sent in the scenario of
test_functional_3_trials, which was not the case out of the CI and
defeated the tests related to the sent emails, which had to be fixed.