Fix double payment of contract-related issues
s2s_do_transaction recreates a payment although created just before
being called in this module. This is because the server action in
product_rental if trigger by the transaction state update and will
create a payment when the transaction is related to a contract (which
is not the case in present module's unit-tests but can easily be
simulated by modifying the server action's code).