Fix phone missing in Colissimo delivery data for non-FR partners
Colissimo used to refuse phone numbers from out of France, so they were ignored.
However a check of the presence of a phone number was just added, so it crashed
for every non french partner.

In this patch, we always parse the phone number of the partner (according to its
country, hoping they match...) and add them (mobile + fix phones) to the delivery
data.