Fix and refactor fees in case of device scrapping
Compensation is now detected based on stock moves, not using
the dedicated project: this way, partial device breakage is
not mode a false positive scrap.

The amount of the compensation is higher than the standard
device price when fees + purchase price are, which is a fix
compared to previous implementation.