Fix portal sale order line details
product_rental was adding a deposit precision which was displayed too
late because of the way the same template was edited in the commown module.

The commown module's modifications were split into two parts to fix this:
one for the product name (after which the deposit mention is displayed),
one before the end of the parent block (using the inside position).

Also the modal toggle button text has been updated to make it clearer
that it can be used to display product details.