Fix sale order name no more updatable through the web UI
The sale order name field was set as computed so was no more editable using the web UI.
Adding an inverse function that does nothing fixes it.

The alternative (using onchange + override create and update/write) is much more verbose
and more difficult to maintain, so I chose this hack.