Fix a crash when generating the last invoice of a contract
When the contract end date is not exactly an invoice generation date,
the previous implementation of the payin bank label generation was
crashing badly, because it was using the recurring next date of the
contract, which becomes falsy at the end of the contract.

It is now using the contract lines' last invoiced date, which is
correctly set by the contract module.