Re-fix the crash when generating the last invoice of an auto-pay contract
Previous implementation shifted the bank label last date by one day.
This was due to a bad interpretation of the test data: a date_end set on
the contract line of the test was trimming the last invoiced date.

The new implementation now uses the contract line last invoiced date,
which is dedicated to the purpose. It uses the invoice lines to get to
the contract lines to avoid wrong usage of non-invoiced contract lines
(for instance because of a zero-computed quantity or such).