Fix fees computation in the "fix" fees_type case
The returned result was always None because of a typo in the fees_type
expected value (requires a unittest).

Moreover, contractual invoices do not help for fix fees amount as the
invoice amount is not used (and is problematic for the yearly
recurrence), so we simply use the period duration to compute the fix
fees.