Do not call the payment_transaction controller from payment_slimpay_transaction
The payment_transaction is indeed called already from the client, then
the acquirer button is renderedi, then the payment_slimpay_transaction
controller is called.

This simplifies the code, which only has to get the last created
transaction and do his job: pay the transaction using a s2s call if a
token is given and return the payment validate page URL, or return a
Slimpay payment approval URL to redirect the user to the Slimpay website
and pay.

The tests had to be completed to account for the extra needed step, but
are now much more representative of the real user navigation.