Fix SMS reminder for support
The `action_send_issue_reminder_sms` action code was different from the
project.task `send_sms_reminder` method and the tests were commented-out
(probably since the v12 upgrade). The code difference matters as the
commented-out tests were not green, and they are now.

Note the SMS language has been forced in the context to the project.task
partner, as it was in the action's code.