Move test code from payment_slimpay_issue to commown_payment_slimpay_issue
This test code was protecting against reverse dependency modules
installation, typically in the CI, which breaks the
payment_slimpay_issue module's tests.

This is now in the right order: the modules that modify the expected
test behaviour of their dependency are in charge of keeping keeping
their tests green.

Of course, an even better behaviour would be that the CI do not install
all modules before launching their tests but launch them in their own
installed environment (like we do locally).