Base urban mine upon project instead of CRM and use 2 different campaigns
First campaign is the same as previous one: buy the phone and give a
coupon. Second campaign only sends a coupon, with a higher reduction.

The customer has to choose the campaign, so there is an extract stage
to ask if necessary.

Also add phone as required in the registration web page, as Colissimo
requires it. The registration page is now updatable as it should not
be modified too much in the UI.