Fix custom module dependencies
commown_self_troubleshooting now depends on commown_devices, not the other way around.
This requires moving the contract termination project creation to a lower level module
as commown_devices requires it.
It also simplifies the implementation of the displayable_description method of the
contract entity type, which can be entirely implemented in the commown_sel_troubleshooting
module instead of partly in commown_devices.