Fix fees period superposition
The until_date of a computation is included: a rent payment at this date
also generates fees. But to ease period from_date and to_date interpretation,
the last computation period to_date will be until_date + 1 day, so that all
rental payment which date satisfies from_date <= date < to_date is considered
in the period.

This fixes a bug of payments at the end of a period being attributed to 2
periods.