When forecasting rental fees, store them monthly to easy graph analysis
When the fees computation date is in the future, the objective is likely
to analyze the fees detailed results. This is much easier graphically,
in which case we assume that the fees will be invoiced monthly and not
at the end of a fees period, which is meaningless for the end user. This
is why we split the fees period into months and store that in the
computation details table.