Add new models for self troubleshooting categories and items
The former is used to group items together in a named section of their
portal in a specific order.

The latter can represent an ST page as before, but also an external link
to a user documentation, like, e.g., Commown's discourse. It will make
it possible to mix links with today's ST pages and order them all at will.