No more ignore some invoices in the semi-automated payment issue handling
Remove slimpay_payment_issue_process_automatically override that used to return
true only for contract-issued invoices.
This was a problem for merged invoices, which were to be handled manually, but
we choose here to handle automatically all cases Slimpay incidents for which an
invoice could be associated with: there is likely no situation, if any, where
this would cause a problem.