Fix a crash when creating a contract-line specific discount
More precisely, the crash occurs when clicking in the 'Replace discount line'
field of the contract.discount.line creation form.

The crash was caused by the domain being empty (false in fact), because the
_compute_replace_discount_line_id_domain method was not called on creation.
To trigger its call, we set default_contract_line_id in the context of the
specific_discount_line_ids field so that the contract_line_id changes when
the discount line is created.