Stabilize a test which was crashing at end of month
The test checks contract forecasts' behaviour, which depends on the date
the test is run. The problem is that contract recurring next date field
is not that predictable as its behaviour at end of month dates is shaky.
We circumvent this problem by choosing a date slightly in the past and
not close to the end of a month.