Do not change sale line name value when not necessary
Previous implementation was using a compute that was also triggered in
the UI thanks to a hack (a no-op inverse method on the field).

This one handles the 2 cases (UI change and other updates, from shop for
instance) using 2 different methods instead of trying to use a single:
- an onchange method for backoffice changes before save
- a write method for backoffice save and shop updates

A simple private method factors out the common code, but it is much
simpler and less hackish this way.