Move contract.line's contract_template_line_id field def from contract_variable_discount to product_rental
This is where it should have gone from the beginning, and it will be
used in an upcoming contract method to determine the "main rental
service" soon.

Note we also removed the discount_date_units class attribute that was
not used anywhere.

Last note: contract_variable_discount test had to be changed because of
the product_rental new dependency that triggers the inverse field dates
computations and lead to an incorrect recurring_next_date value before
the change.