Fix wrong purchase price computation in rental fees
When a compensation occurs for a given device, its purchase price (reduced as
agreed in the contract) is computed using the invoices attached to the PO.
Before the patch, all invoiced products were used to compute the price (as
their mean price), which was wrong as different products may sometimes be
bought and related to the initial PO line (which is bad but occurs when the
user modifies the po-generated invoice line and changes the product).
We now filter by product to avoid a wrong price computation for the device.