Reworking step-ship-module in common_steps.xml
Since we want to add this step to all our ST pages, we do need to reword this part of the text.
Therefore the customer know why he should verify his shipping informations.
Then we would remove the new_module variable since it is now reccurent.
The module / device isn't longer explicit in the form itself for customer,
but we'll always specify it in our shipping confirmation mail.