Fix incorrect domains of project tasks product and lot fields
They were using domain methods not called on existing tasks. They are
now using not stored computed dedicated domain fields which make it
easier to debug and more reliable.

Note the test method dedicated to the UI (prepare_ui) had to be enriched
by a mean to evaluate the domain attributes defined in the view.