Fix Send as invoice recipient action for contracts
All the contracts which main partner is the parent of the invoice partner to become
the recipient of invoice emails must have their invoice_partner_id modified by the
action (a button presently).

Before the patch, the selected contracts were the one which invoice partner was the
parent of the recipient contact.