Forecast cooperative campaign discounts as their last computation
To this end, we check for each applicable coop campaign discount if it was applied on
last invoice and keep that result for contract forecasts.

Note this last-invoice test is fragile for now as the data model does not cleanly store
which discounts were applied on each invoice line. This will be improved in an upcoming
commit.