Add a domain to crm lead contract field
This makes choosing a contract less error prone.
Note the contract_id field comes from commown_lead_risk_analysis but
commercial_partner_id from commown_shipping, so we chose to add the
domain here to avoid adding to much dependencies and because it is not
essential but only a UI helper.