Add ability to use the contract line start as variable discount ref date
Before the patch, the reference date was necessarily the contract's one,
now it is possible to choose between the contract's and the line
contract's.

A data migration script is added to make sure the DB values correspond
to the data model, BUT there is a semantic change for the non
"contract:" prefixed discount date reference values so the discount
dates will change for some contract lines. This is addressed using an
external script.