Fix the invoice merge preferences action on new token signature
The algorithm permitted invoice merge dates in the past. It was too complex too.
Now we just use the more recently used preferences (= the one that generated the
more recent invoice merge date) and increment the date according to the recurring
parameters to make sure it is in the future.