Granting employees portal access is now delegated to a specific wizard
This wizard is used to warn the user this operation has security
implications.

It also asks the user to input its password and checks it on wizard
creation (without storing it in the db, for the sake of security).

The wizard can be executed on employees with any portal status as it
filters out those which have already been granted the portal access.