Make it possible for customers to manage their company's users
B2B customer users can now create, update and deactivate partners as soon as:
- they belong to a customer "admin" group
- those partners are in the same company than them

The customer admins can create a team hierarchy in their company and
attribute them to collegues.

They also can grant access to the portal to their collegues, which
requires to enter their password for the sake of security.

The design choice to make customers directly manipulate partners was
made compared to:
- using a dedicated model with fields synchronized with partners to
  avoid a difficult management of the life cycle of such entities
  compared to partners (when should we create/ unlink them?) and the
  data migration
- using a dedicated entity with delegation inheritance, for the same
  data migration difficulty

The main drawbacks are:
- security: access to other partners is a risk mitigated by the unit
  tests and the CI
- performance: added rules might heart performance, but this is limited
  to portal users and should not heart in the short and medium terms