Test fixes
In commown_devices tests, the po was using the default picking_type_id
leading to devices arriving in the sale stock.

In rental_fees tests, the repackage step was missing but was not
crashing because the repair location was a son of the default
send_device origin location (which was stock.stock_location_stock at
that time).