Prevent customers from removing their own admin role
Also move the "short" tag option into a dedicated kanban widget. This
widget does not handle clicks anymore.

Preventing customers from removing their own admin role is done using a
"readonly" role computed field (which is given current partner_id as a
context). The value of this field is fetched in the widget and used in
coordination with the global edition or view mode to display the removal
icon and to prevent keyboard removal (or not).