Remove the assumption parent_id is not sent when a customer manager creates contacts
... so instead of checking parent_id is set to the permitted value (=the
company of current user), override its value instead.

This is far more reliable and future proof, as the parent_id field may
be sent with False as a value instead of not sent at all depending on
view configuration.

This impacts the contact import feature, as we also need to override
the parent_id field's value when updating a contact, which was unless
before.