Add ability to send devices part of a sale service from a project task
For rental, we use crm leads to first study the risks and validate the
rental.

For sale with services, we use project tasks to prepare the devices, as
we do not need a risk analysis once the sale is confirmed.

The wizard to send the devices is now common to both use cases.