Fix wrong `all_products` permissions for customer picking wizard view
The view with the permission restriction is the one to be
returned as Odoo manages to apply the restriction only when
applicable (=when current user is in the restriction group).

Before the patch we were returning the view without group
restriction, which obviously gives the same result for all
users, which is wrong.