Added customer admin attributes skip if the env. superuser tag is positive
In the res.partner write method in the base model, there's a _fields_sync method, which writes "commercial fields" (returned by the res.partner method _commercial_fields) if the parent_id value is changed, as a superuser (with the sudo() method).
Since the modified fields in _commercial_fields are not directly modifiable by a customer admin, instead of adding these commercial fields to the allowed customer admin attributes, we instead skip the check if the superuser environnement variable is False before checking.