Added sudo() for read operation on ir.module.category
When a user in the group_customer_admin tries to fetch the groups, it fails as the domain reads the `category_id` of `res.group`, a `ir.module.category` record which can only be read by internal users.
So, using the sudo() method permits the user to temporarily read this field to perform the action.