Added export-group inheritance to customer admin group
Starting from Odoo 14.0, users have to be included in the `base.group_allow_export` to be able to export data, and to be able to see the Export action in the Actions menu.
Since customer admins should be allowed to export customer data, we grant them access to this group.

This group is not limited to internal users so we can share it with portal users, like the customer admins.