Added /web route override in customer_manager_base
In the `web` and `portal` modules, the `/web` route controllers contain conditions checking whether the user trying to access the `/web` domain is an internal user, then either redirects or forbids access.
This is problematic since this module's functionnality relies on giving a limited access to the `/web` side of the Odoo app, to be able to perform actions on users limited to their company.
So, we replicated the code in `odoo/addons/web/controllers/web.py` but added a check for customer admins to allow access to /web

NOTE: The way permissions are set, customer admins should only be able to access the records linked to the customer_manager_base module and its children, but this is still granting access to the backend of the Odoo app, so model access rights need to be limited and justified.