Moved IrHttp override in customer_manager_base
Since this model is concerned with setting a session variable linked to the `group_customer_admin` group, and this group has been moved to the customer_manager_base module, we moved this model to the corresponding module.

NOTE : In the moved test, we only add the `group_customer_admin` group but no company, since in the current iteration of the test, a company is not required to obtain the `group_customer_admin` group.
A feature which could be added is an override of the `res.users` model that checks whether the user is B2B before being able to get that role - however, this is not the goal of this commit.