Execute tracking instructions after slimpay issue task creation
This aims to allow the attribute tracking actions that follow to take
place in the same transaction.

The task creation inhibits further attribute tracking actions for the
whole transaction by calling mail_thread._track_discard.

This is not appropriate here as we want an email to be sent when the
task is moved to a stage with a mail_template_id, which is done through
the tracking mechanism in Odoo 16.0.