Deleted references to removed scrap location
The stock module changed the way various locations (incl. scrap) are generated - rather than having a general scrap location, the module creates a location for each company in the environnement.
However, we don't need to set a new default scrap location, since the stock.scrap creation sets a default scrap location, which is either the default company or the environnement company.