Fix and test payment fees
We now use raise_if_not_found=False parameter of env.ref to return None when
the concerned payment fees product has been deleted (which can be a way to
handle them manually).

Before the patch, this was done correctly only for the bank (supplier) fees.
This is now also the case for our management fees. Both now use the same
method.