[test] Removed action ref. autocomplete in _trigger_obsolescence
Beforehand, to allow shorter references to actions in tests, mainly in the commown module, which then auto-completes with the start of reference for an action declared in the commown module.
However, this makes this part out of place in the payment_token_uniquify module, reducing its readability.

So, we removed it for visibility