Modified view to not depend on moved field
Previously, we used the `journal_id` field of `payment.acquirer` to concatenate the obsolescence token field right after it.
However, the `journal_id` field has been migrated from the `payment` module to the `account_payment` module, meaning we can't use it without changing the dependency (and adding more modules to download), we instead directly inherit the original view to add the obsolescence tokens field at the same place.
Also, since the payment_followup group is invisible by default, we add an attribute change to invisible to insure it'll be visible.

NOTE: since there are no menuitems linked with actions in payment to access payment.provider records, installing account_payment locally is still required to check if the field is correctly added or not, because it's where the menuitem that allows frontend (browser) access to the provider records.