[test] Added test for session expiry
In the `web/controllers/home.py` controller (which we replicated), we check the validity of the session with the `check_session` method, which compares the expected session_token with the given session_token if they match.
Modifying directly the session_token is not possible : before the controller call, the session validity is checked, so modifying the token already invalidates the session, which gives an anonymous session.

To circumvent this, we created a mock which returns True on the first call (before the controller, to not invalidate the session) but False on subsequent calls (normally, only the controller call), so we can easily match the condition.*

NOTE: For the calls list we use to count the number of calls, we make use of a dangerous default value in order to broaden the scope of the calls list to both the test method and the test function, so the list object can be modified and have its changes kept in memory in both scopes.