Improve the appearance and ergonomy of the contract view
... by grouping related fields together in groups.

Added fields' position was (before present patch) based on other fields position.
This is changed here because fields can move around from an Odoo version to another
so this is not a good criterion. A better one is chosen here: existing groups.

This will hopefully making the maintenance less of a burden...