[test] Added test for _render_qweb_pdf rendering py3o reports
The _render_qweb_pdf is a method used in some views to render reports that are initially declared as qweb-pdf, but that we overload as py3o reports. So, we overloaded the _render_qweb_pdf method in order to correctly render the now-py3o reports through the _render_py3o method (see custom_report/models/ir_actions_report.py).
To ensure that the overloaded method doesn't encounter any issues for both qweb-pdf and py3o reports, and that the returned files are pdf's, we wrote this test testing both the base behavior (qweb-pdf) and the introduced alternative behavior (py3o).