[test] Removed test cursor for product.product creation
After the with block was closed, while the product.product record was created, it couldn't be found in database queries (for instance, in the create_contract method, trying to create a contract line with the created product would fail by saying that the product's foreign key doesn't exist in the database table).
However, when creating the product outside of the test cursor, there are no issues in the create_contract method, when creating a contract and its lines.