[test] Removed date_end computation after contract creation
In the creation of the contract, if a date_end value is given, then the date_end field of the contract lines are matched with the contract's value. However, using _compute_date_end sets the date_end of both the contract and its lines to False.
So removing it leaves the date_end value we pass it.