Adapt to change in _compute_price_rule method.
There are quite a lot of changes since the function used to take a tuple
of product/partner/quantity as an argument and now takes a recordset of
products and a quantity.
When doing tests in the interface we noticed that most of the time it is
called for a single product. Nontheless we made an implementation that
can take a recordset of products as argument.
To read with extra caution.