Migrated SmartWizard lib to version 6
In Odoo 16, the SmartWizard we use for the self-troubleshooting forms (version 5) started to show some visual glitches. For instance, the progress navbar's transition weren't as smooth (the blocks colors appeared with some delay).
So, we decided to migrate the library to its version 6, released in 2022.

Notes about some changes :
- the state of a step is no longer a unique value, but rather a collection of states that can be activated at once (for instance : a step can both be 'done' (visited) and 'disable', and there are no 'enable' state since not having 'disable' qualifies as being enabled.
So, we have to remove a given state with 'unsetState', and set it with 'setState'