Query merged invoices in SQL with a clause on analytic_distribution
This field is stored as jsonb-typed postgres column, which is not
yet well supported by Odoo and not easily queried with a domain.

This is why direct SQL is used here.
Note that the field is indexed in postgres, so the query is fast
(and this is important because the table is huge).