[test] Added account reloading for expanse account
In a local installation, the environnement company (base.main_company) doesn't match the company for which the l10n_fr records are created. This is mainly problematic for the expense account of the Rental Fairphone product template, which crashes the account.move post action.
So, to pass tests in a local installation, we reload the l10n_fr records for the base company.