Added invoice_date field set in account.move creation
To be able to perform the post action on a purchase invoice, it's required that the invoice_date is set. Since the related purchase.order's purchase date is also set to the day of the execution of the method, we also set the invoice date to the current date on which the invoice is created.