Remade Urban Mine form with Odoo 16 page editor
Since the website_form_builder module has been fused in the website module, the format of built forms has been changed, so some values weren't correctly parsed when creating a partner, and the form itself wasn't recognized by the editor, making its modification difficult.
So, we reimplemented the form with the new editor.

We also took the opportunity to change the 'supplier' field to 'supplier_rank', with a base rank of 1.