[test] Added OCA/website PR to test sources
With our PR to add captcha API customization to OCA/website (https://github.com/OCA/website/pull/1111), we added fields in the website model, which we use in the urban_mine module.
So, we add this source in our test requirements in order to fetch the code in the GitHub test job.