[test] Added Expl. tag test
Through action_registration_payment action, we add a tag named "EXPL" to created invoices, in order to count the invoice as part of the company's spendings with the urban mine.
Currently, the tag was not added through any module data, but directly through the web UI.
So, we create a demo tag with the same name (EXPL) to test whether the tag was correctly added to the invoice record.