Added an empty dict. return for mandate_doc_to_repr
If the method can't find a partner to link the mandate to, it doesn't return any data, which is then considered a Falsy (None) in get_all_mandates_repr. In order to clarify that possible end point for the method, we add a return with an empty dict, which is also falsy.