[test] Improve cov.: Added no cover flag on MissingError catch
The try/except block tries to catch a MissingError exception, which is thrown in the Odoo code when trying to access a record that's been deleted.

The only accessed records in the method are a partner record and its related payment token record (which can be non-existant). Since the start of the method already verifies whether the partner exists, and writing on a non-existant token is not forbidden, the only way for this exception to be thrown would be if a concurrent writing process were to delete that record after the start of the method, which is tricky to test reliably.
So, we add this flag on this exception.

For the Exception branch, while it's more common to encounter issues (for instance, a ValueError trying to access a undefined dict value), we do not consider it interesting.