Added superuser flag for write/create project.project operations
Since Odoo 13.0, sudo() no longer performs a given access or action with the superuser, but instead changes an environnement flag - su -, which can be read to see if permissions need to be bypassed.
As such, to skip any access right verifications or other behavior if the su flag is active, we changed the conditions to check access rights or value setting.