Changed mail.message model used for the actions
In Odoo v12.0, the mail module inherited the mail.message model to add a method and overload the _search method, but no longer overloads it in Odoo v16.0.
Since it was an inheritence however, it's more fitting to directly assign the original declaration of the mail.message model from mail.