[test] Added flush/precommit to tests related to stage changes
Changing the stage_id of a given task is not enough to trigger a message post of the stage change.
Indeed, stage changes messages are only sent during the commit, with the precommit script.

So, to force a post of the stage change log, we add flushes to trigger the posting of the message.