[view] Fix website_sale_b2b website settings view and adapt commown's accordingly
In website_sale_b2b, form fields were added outside a group so their labels were
not visible. The tab where they were added was not suitable for them, so we added
a dedicated one.

In commown module, we take advantage of this tab to add a new group with our field.