Create server actions to generate a PDF file of coupons
While there's an already existing action in the Odoo v12 instance, it is only declared in the database, which can mean its presence is uncertain.
So, we add a permanent action in the code dedicated to this action.

Note: since the program is written to write 65 coupons (it loops on the width and height of a pdf file which leads to 65 coupons), we named the action accordingly.