Move pdfminer to commown_website_sale_coupon and pin its version
The attachment_indexation module requires pdfminer and advices to
install pdfminer.six (which is the maintained fork of pdfminer).

However, recent pdfminer.six version require cryptography>=36.0.0 which
is not compatible with pyOpenSSL 20.0.1 which is installed on Ubuntu
22.04 (generates CI error: module 'lib' has no attribute
'X509_V_FLAG_NOTIFY_POLICY', see PR 8769 of pyca/cryptography on
github).

So we pin pdfminer.six version to 20220319 which is the apt version in
ubuntu 22.04.