Modified file attribue check in account details endpoint
Beforehand, the controller tried to access the filename attribue of the given object. However, if the given variable does not possess a filetype attribute (wrong type of variable), this might lead to an AttribueError being thrown.
As such, we add an operation checking whether the given value truly has a filename attribue and return that value if it exists.